+++ CHANGELOG.md
@@ -11,6 +11,7 @@ The format is based on [Keep a Changelog][keepachangelog] and this project adher
 - Done migrate to typescript [#116]
 - New settings storage implementation [#119]
 - Indent size now equals `2`
+- Settings allows to trigger custom handler on settings changes
 
 [#116]:https://github.com/tarampampam/random-user-agent/issues/116
 [#116]:https://github.com/tarampampam/random-user-agent/issues/119